test: Fix exccluded tileset-proxy test
