+++ test/testdata/multivec-data.js
@@ -1,5 +1,4 @@
-// @ts-nocheck
-export const multivecTileData = {
+export const multivecTileData = /** @type {const} */ ({
   dtype: 'float16',
   maxNonZero: 52672,
   minNonZero: 3150,
@@ -14,9 +13,9 @@ export const multivecTileData = {
       '[8328,7776,13240,11104,8600,15760,13480,10560,10208,10840,9360,9672,11296,13392,18896,19568,20880,28048,33600,39840,30112,33952,27328,15144,27664,31376,31120,25104,21808,22352,30320,31744,25472,28976,30496,32464,23856,18256,8984,31968,25472,21104,17040,13016,23328,25312,16592,14440,12896,15544,19040,23152,14288,17392,19104,18992,12808,19072,25232,22432,16992,18912,28576,25104,22272,12888,16336,15568,16432,21056,21472,11760,12376,10528,14600,10072,9952,10032,11664,13224,13168,11656,11080,10712,11760,19120,15968,13776,12264,15080,15376,14216,19344,20448,18496,17936,12880,14472,17600,16544,15152,18016,20048,17776,20416,16184,21408,13920,21504,21072,21984,11056,12056,19376,29568,30448,24624,26832,27488,15288,16944,21872,26352,39328,33280,16608,23136,24000,25168,22992,20544,28176,22576,22544,26496,28048,22160,18304,17408,17632,19392,17360,22416,15960,15736,20048,15336,20976,30384,26096,31072,21360,21648,16288,21648,15856,17200,18080,15824,10136,10760,14792,11056,9360,9752,8336,10000,11056,10232,16080,14504,15656,15808,13912,13928,15720,17488,18128,14456,11240,12848,8052,10072,13960,10600,15992,20960,15304,24048,28832,21504,20784,20096,15112,13280,21472,22624,19280,20944,12384,8896,8060,8240,7088,34368,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,23776,14160,24592,26576,26048,21872,25968,26288,26912,27376,23552,21616,21744,15040,18112,19648,23104,9680,11544,12344,15128,13592,12256,12216,16008,14856,13744,11464,15440,12080,14432,14912,14464,13808,13424,13592,14448,9872,5260,15144,15848,21248,26480,14560,16832,12280,18864,17872,19280,19248,17552,16400,20480,20864,17936,15568,14472,18944,15880,18080,20784,18464,16344,18384,17744,19296,15696,15928,18576,18960,21440,20816,23264,21872,22992,24416,23472,24448,24016,22368,18672,19808,23600,20784,20720,22336,21360,23312,22592,22448,17536,18272,21040,19920,19872,18496,21232,20944,22128,17680,18768,15496,24784,19952,13064,9280,15088,9928,17264,16720,13992,17216,18112,20960,13480,17488,16464,10808,13616,22320,20448,13952,10232,13696,11800,9128,14392,13872,17248,16304,15816,14536,12232,14208,16720,14368,23392,16432,16016,16336,18640,19472,15752,20016,19504,19168,18080,15968,20592,14872,13864,20464,13944,21824,11616,14256,15752,16880,14120,20048,20992,17136,18256,19952,19872,21712,20224,21664,20144,17520,16024,21072,16168,21360,19040,22464,24384,16736,13704,14896,15816,6104,13048,20192,21280,17280,14440,18272,16368,17232,12696,12272,13272,15144,18608,17024,15640,13184,13304,15896,16896,15984,16296,9640,15448,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,11488,8504,14664,13576,11528,17328,16304,13776,13528,12304,13360,12056,13112,13456,13880,15440,19984,15656,20432,28880,21104,22480,20368,11496,17152,25904,18768,16168,14344,16800,16992,17776,13256,17088,20576,21632,18320,10544,6400,21472,16640,17936,16072,12096,16104,15904,11576,11472,12152,13600,13456,17088,12728,11960,14800,14568,12744,17344,17088,17232,15320,15304,20240,18656,17952,12160,10104,11472,10632,16832,18880,12304,12832,11800,14328,10704,10520,11640,10352,11200,10688,10848,12760,10360,11824,18016,14048,13640,13168,14568,12304,13456,18768,16208,16688,15472,13480,13504,15256,14784,13072,13016,19440,14208,14464,10944,19552,12192,18304,16120,15192,11360,11424,19648,23776,27216,21568,14072,24912,13832,15504,14184,27456,37856,28384,10592,15736,15304,18400,21456,13944,19440,12744,16336,17904,16248,21984,13896,13648,14272,14912,14976,14856,18032,14592,18128,14784,16672,27344,15888,21136,20592,18016,14632,14544,10368,12496,14928,11896,11008,9456,11488,9616,9064,9560,9000,9248,10432,10688,13120,12328,14328,12696,11888,10944,11832,15648,13360,12104,10592,11000,6332,11792,13384,11328,13088,14712,15808,18208,20192,12192,11824,13208,13296,12616,19696,15296,19920,16248,9808,9216,9240,9344,6508,27456,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,12752,9936,17792,11488,15664,20032,21088,10728,11048,10272,9344,9472,10192,10304,12752,16672,25136,19760,20864,21664,22688,35872,33312,13392,29200,37472,24368,19856,21264,13768,20304,26592,15848,20736,28544,30176,22112,13272,10008,37152,19280,24912,19984,16248,27248,20752,14936,12256,10808,10984,12440,18208,14896,9680,9584,11744,10808,23936,30048,26992,19424,24656,29600,25600,25136,12696,13288,13104,11792,18448,13632,11208,10352,10112,15584,9736,9032,9336,9248,10224,9784,9616,8848,9472,10544,14984,10296,11120,16168,12752,12440,18816,25216,14200,12336,12144,10248,10184,15944,11312,11272,16496,26784,11912,18544,12096,29008,14368,28032,24144,14864,14408,15152,16896,34784,46336,27040,19008,25200,13488,22848,16800,21824,46016,36928,12984,13768,16752,26224,29728,14560,23776,14680,17328,24224,23328,28080,13088,11928,11000,11640,12136,15456,12216,11824,16400,12168,15040,36832,24080,26144,32672,21120,16448,13664,13648,20432,17488,11000,10072,9064,9608,9448,9136,8240,8632,9184,8824,9000,9992,9816,9904,9712,9960,11448,15376,20544,17984,10144,9608,10808,6076,19648,12920,10952,17776,18000,20688,25456,33152,14712,13968,14440,16752,15320,22944,18112,29696,25344,13984,16208,9448,9896,4332,37792,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10552,6888,12960,11352,9952,14760,14424,14000,12864,11320,11304,11144,10768,11456,15496,15712,16240,14920,18480,25456,26144,22320,19616,14776,18688,21920,22288,25456,18672,18112,23968,18960,14680,17184,18848,19232,18496,12960,8704,23344,21600,14928,13504,10080,22672,23024,12672,12280,12904,12448,14464,18672,16448,17680,18368,16592,12480,15448,15432,16208,13504,17232,18784,17104,13336,12032,9464,9160,10504,17168,18432,14408,14104,13072,14616,12152,11616,10248,10224,8832,8584,9888,11656,8880,8816,18080,14904,14944,14320,17456,14992,12584,16512,18800,16232,18416,15208,15952,18368,18352,15592,15008,18144,13400,13256,8952,21184,9824,18368,20720,17568,8992,9744,16880,21344,18880,16120,19312,21088,13904,15824,12616,16688,21248,21456,9080,16144,16088,16088,16240,11136,19504,12048,14672,19968,14688,16112,16800,16192,13784,15872,15856,18032,15160,12440,13192,11896,10944,20624,21616,23552,16512,17200,17408,21104,11136,13400,15136,8520,7696,7020,8424,7944,7848,7448,6452,7600,9416,7728,11600,18960,18112,16144,12448,14016,15112,14320,14752,7720,8008,8288,5116,10824,16272,14064,18464,19136,14080,17120,17120,18880,18464,19472,17728,14792,16992,16912,13824,14704,9496,7720,6524,5984,4232,18496,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,30192,20720,28352,24720,26080,23392,27040,24432,23584,23696,21904,20640,23504,22960,23984,25664,28496,18176,17120,19776,20176,24976,21200,22480,22896,21968,19744,18144,22624,18944,21904,22496,18976,18912,21008,20768,21088,14808,8976,21440,20432,25184,24416,18368,22752,19056,18864,18880,18960,19712,18880,19520,20464,22032,19328,18784,18096,22272,22432,24272,26560,23536,23552,25920,25344,21728,17952,17600,20112,21840,24176,20080,21104,21904,25648,20576,21424,21840,22112,20560,20560,22624,25552,21312,23024,27648,26800,26960,29056,25232,20976,20768,27024,20240,23248,22384,23872,23472,24640,22048,20192,19392,28208,21936,17808,13704,22048,18432,23600,22432,18736,22640,22560,27184,23904,25200,23280,17696,19904,27072,26720,22112,17232,20944,17232,15000,21664,20720,23328,22688,20496,19728,17456,20704,21872,23792,29568,22224,21024,21248,21392,24256,23728,26112,23168,22528,24112,23824,24992,23984,23392,27392,22688,24928,19872,23584,24208,21696,18128,20784,20432,21056,19184,20304,21680,22080,21456,22128,20848,21888,22448,23792,21232,22000,20624,22000,26800,21536,20896,22208,21792,15488,21616,26080,27584,24784,21952,25040,23040,25904,18592,17904,18288,21552,23840,23104,23280,21904,20368,19392,18448,17552,18800,12952,22880,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,23248,18160,23952,22064,22656,23136,24416,22432,21920,21872,21312,21408,22080,21360,22448,23168,24864,17472,16992,19312,19440,21344,18048,20000,21424,21568,18448,17872,21296,19312,20688,20656,18704,18960,19824,19888,20704,14976,9304,20768,19648,22160,21984,18480,21104,17952,19280,18448,18528,17632,18272,19264,19600,19872,19152,17664,18224,21952,21072,22176,23200,21824,21584,22464,22368,21616,18320,18816,20224,22192,23056,21088,20688,20432,22704,20624,20512,21008,21856,20576,19808,20320,22544,20816,21424,23824,22192,22640,23104,21664,19616,20896,23376,20528,21440,21072,21200,20832,21776,19472,19552,17888,24560,20624,17472,14888,19584,16384,22720,21312,19456,21984,22416,22736,21136,23680,20640,17664,20128,22896,22512,20416,17264,20336,17008,14056,19632,19456,21040,21840,19696,19760,17712,19856,21776,21504,24384,21232,20464,20144,20800,21296,21376,22944,21728,20704,20752,21136,22864,20544,20400,24704,21280,22608,18992,21328,21856,20976,18160,20672,20336,19632,19248,19760,20592,20688,21024,21504,20448,21488,21872,22400,20272,20432,19520,20656,22032,19824,18288,18672,20320,14656,18496,22928,23920,22688,20240,22144,21536,23264,18576,18064,18768,20928,21616,21280,21200,20800,19600,19008,17120,17536,18144,12688,20848,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,12856,8464,15880,14440,13768,18800,17296,14168,14656,13472,13992,14848,14560,17408,17008,18240,20944,23600,26656,27040,27408,25760,25936,15360,27552,28368,27248,20288,21392,22576,27760,30928,24880,24896,24336,26304,23536,16752,7572,27024,22176,20128,18320,15880,23968,25920,22272,19104,17776,18464,21184,24320,20320,24912,23408,21872,16120,20720,22288,21248,19504,17776,26400,22624,22672,16176,17648,18656,22976,21632,19216,14616,16264,14344,16944,14368,13792,16448,17920,17440,19920,16072,14288,14376,15976,18928,15856,14792,17296,16800,16088,15776,18752,18864,21440,20032,15528,15640,18192,17328,16176,15504,23760,16024,20688,17504,25296,11848,20288,20704,20304,14256,16128,21808,27568,32768,26464,22176,24592,14424,18336,21840,21696,30576,27504,14456,22640,21792,22720,25264,19296,22672,21520,22128,24848,23104,22512,19184,19312,18416,20176,21120,23088,19936,18544,17136,18112,20336,29904,22928,23728,22032,21488,18400,22912,19312,19504,20384,17760,15224,14224,16576,14392,14664,13488,13000,14448,16768,14480,19184,18384,16240,17584,15608,17200,21184,18560,19632,16048,17584,16144,9744,12608,17840,16024,20832,20336,17600,23456,28080,24432,22592,20336,20656,15088,18768,20720,25344,23408,17760,11616,8248,8528,5732,20192,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,11752,7272,11904,10256,9904,14336,12720,10472,10704,9352,9344,9832,9984,12360,14320,15048,17808,19872,19696,22880,19936,24304,20560,12536,21296,24736,19008,18336,15912,18992,21280,22112,18080,20048,21456,21648,18720,11016,6352,19328,15680,14264,14256,8808,16096,15656,12552,10888,10384,12024,13760,16400,12728,13664,12880,11680,10000,13488,15392,14848,13720,15048,18256,18544,19040,12544,12608,12584,14264,15856,17712,11048,10792,9768,11992,10968,9904,10040,11768,11528,12016,10512,10328,9456,11128,13272,9672,10272,10384,10200,10768,12128,14032,14976,14016,13336,10264,10720,11824,10720,10432,11992,16032,13776,15008,10864,17376,10536,17392,16136,15904,10432,12320,18352,21760,24608,19344,18256,17456,14400,14064,14280,16896,27984,20640,10384,14576,15072,19008,17552,14072,16752,15432,16112,17872,15104,16128,13552,15008,14864,15888,15488,14832,14096,13240,12680,11648,13920,17552,17264,18272,18320,19168,14888,16152,14000,15024,14368,14280,10864,10240,11000,10464,10184,9856,9248,9960,11280,10424,11920,11072,10312,10544,10880,12200,13152,12376,12440,10520,10280,10848,6540,9200,14544,12208,14872,14888,16264,19472,19424,14856,13760,13640,11792,11248,15584,15752,18352,18272,12344,9616,7828,8024,5820,19840,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10744,9168,18208,9760,14896,20112,21024,9752,9544,8880,8368,8992,9328,8920,11624,16432,26224,18272,20656,20208,23008,38880,35808,11496,32400,39968,23648,19392,22384,13296,20848,25936,15016,22448,26736,28304,23152,12520,8888,40032,16944,26416,20080,15952,25728,19888,12768,10304,9672,9640,11112,17408,14000,8768,8424,10288,9592,25936,31408,25728,19328,24848,28448,28208,25600,11720,12056,11400,10824,18608,13192,9536,9176,8824,15008,8808,8296,8544,8424,9272,8928,9392,8808,8880,9064,15600,9432,10048,14912,11104,10832,16832,27568,12176,10568,10744,9512,9920,13608,9768,9480,16640,26544,10296,17392,10704,30736,16056,31376,22928,13352,15232,11096,19168,36064,49600,34272,15280,25664,12256,22496,18688,20752,52672,41536,9368,12408,15200,25968,37152,13688,22176,13960,16864,26048,20832,32384,12256,10856,9696,10520,12000,14128,11080,11096,15432,10576,16912,38336,21616,26272,34976,23168,15704,12920,11464,20832,18400,10152,8896,8568,8720,8584,8172,7912,7828,8368,8032,8400,8968,9272,9152,8888,9064,10824,17072,17648,18640,9096,8952,10056,5720,21536,11936,10136,16120,14096,20928,35520,32160,13984,12424,13200,15432,13520,23184,18944,32448,26432,12480,14872,8640,9400,3722,40960,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,11824,10024,16944,12864,13576,23920,22784,10392,11248,10176,9056,8904,9968,10512,12240,16848,26336,19760,19808,21472,22784,43168,28480,12672,32800,37056,25008,17760,21648,12952,19008,28208,15504,24992,25392,28944,23360,12448,9448,38784,20592,23024,19504,19008,25808,20592,14128,12528,10424,11040,12080,19440,14432,9656,9656,11816,10968,27008,26256,29296,20368,26736,26144,29008,25968,12408,13144,12624,11944,18928,13080,11160,10120,9984,15360,9736,8712,8752,9328,9848,9512,8872,8232,9048,11560,13360,9960,10560,17104,12528,12280,19136,26960,14136,12896,12000,9728,9576,16480,11952,11792,19248,25936,13552,16640,11816,25216,18768,32096,23056,16656,12000,13344,18800,36224,44064,32256,19344,19520,19568,23232,16736,22240,48320,37152,15920,13304,14112,29776,30032,13792,23248,14056,17584,23936,26576,28064,13704,11616,11016,11472,11096,16384,11792,13072,14656,12384,16360,34720,26656,23920,35392,22912,17200,12968,14440,21040,17200,11064,9488,8864,9136,9048,8888,7848,8312,9192,8488,8584,9664,9560,9384,9256,9840,10696,17104,21088,16040,9832,9232,10120,5948,20640,12232,12056,18560,16832,20128,28576,34016,14424,13304,13568,17136,14712,24896,20272,26656,28944,13360,17744,10208,9064,4760,39456,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,11752,7272,11904,10256,9904,14336,12720,10472,10704,9352,9344,9832,9984,12360,14320,15048,17808,19872,19696,22880,19936,24304,20560,12536,21296,24736,19008,18336,15912,18992,21280,22112,18080,20048,21456,21648,18720,11016,6352,19328,15680,14264,14256,8808,16096,15656,12552,10888,10384,12024,13760,16400,12728,13664,12880,11680,10000,13488,15392,14848,13720,15048,18256,18544,19040,12544,12608,12584,14264,15856,17712,11048,10792,9768,11992,10968,9904,10040,11768,11528,12016,10512,10328,9456,11128,13272,9672,10272,10384,10200,10768,12128,14032,14976,14016,13336,10264,10720,11824,10720,10432,11992,16032,13776,15008,10864,17376,10536,17392,16136,15904,10432,12320,18352,21760,24608,19344,18256,17456,14400,14064,14280,16896,27984,20640,10384,14576,15072,19008,17552,14072,16752,15432,16112,17872,15104,16128,13552,15008,14864,15888,15488,14832,14096,13240,12680,11648,13920,17552,17264,18272,18320,19168,14888,16152,14000,15024,14368,14280,10864,10240,11000,10464,10184,9856,9248,9960,11280,10424,11920,11072,10312,10544,10880,12200,13152,12376,12440,10520,10280,10848,6540,9200,14544,12208,14872,14888,16264,19472,19424,14856,13760,13640,11792,11248,15584,15752,18352,18272,12344,9616,7828,8024,5820,19840,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,15464,11176,24448,24800,21408,18320,31312,21120,22432,21216,15856,14360,15832,12432,16024,21856,24496,11840,15480,22784,20688,24752,15832,14040,18448,17472,13672,15176,21248,14760,25952,29552,14264,14952,18480,18640,17424,8744,6920,21984,15648,31152,30992,12232,16032,15488,10768,11424,11144,14056,14560,16064,15352,17136,15152,10760,10224,15664,17856,26512,21824,18784,28064,22704,17232,14944,10736,10752,12032,19360,28592,14560,14296,13048,17856,13336,13856,15800,15160,12728,12256,13416,17680,13320,13096,22528,18752,17504,16832,14232,12472,17552,21312,17504,19520,16120,15960,15944,16768,13504,12360,13928,28400,22976,11416,8128,16672,10488,25168,16184,10488,11856,14536,25200,20768,22720,18880,11344,15296,18768,20336,12840,14720,17248,16880,11432,15624,17120,16992,17728,12648,23296,11232,12656,16104,13016,22640,13808,11528,11584,16400,15168,13768,17664,18048,19264,15184,13536,31744,17296,25904,24224,16136,16992,13232,10976,15680,15040,10144,13016,13376,15080,11096,11456,13568,11984,12040,11936,11048,12576,12616,17072,13400,14640,11736,15584,25056,20688,10144,10160,10520,5340,15152,15128,14992,12168,13288,15704,22608,17792,9888,9592,10592,12136,13504,16896,17712,13704,13840,10896,10768,9712,11160,7308,20032,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4516,6064,12944,11848,10064,16048,19184,8824,11320,8952,8824,6584,9816,15504,18784,24544,31728,28368,36608,51872,41120,47360,38272,14888,35392,33728,27568,37216,35424,33568,44416,45792,25072,27072,39296,37440,29584,18592,12432,46528,31296,31808,21424,12752,26032,32080,12448,10168,9264,12200,16040,22400,14080,17776,15088,12104,9800,19424,32384,38592,23088,22112,45888,27584,24576,12520,11376,11520,11704,20944,28016,8424,7756,7068,11064,6196,6868,8036,9312,7496,6268,6848,11104,7820,8592,20480,17120,11064,8568,9232,9552,15664,20736,17984,24224,17296,11920,14104,16088,15208,11392,18960,28288,21808,18272,14872,35488,12768,32624,21520,16352,5784,12856,28256,43616,43328,27584,29312,33472,13504,18848,16008,32448,48736,37664,23648,25440,28288,29392,25824,15192,34848,17904,18336,26416,23792,26528,17008,15344,17360,20032,16864,21872,16024,19360,22944,16752,20096,43328,31008,50656,30528,26032,17904,22672,12040,20352,19184,10784,7200,7268,10912,6348,6648,5636,5156,5004,7268,5996,11064,11576,11872,11984,10720,8792,11592,22672,25520,10840,8272,8960,5464,13536,10840,6904,13240,21456,16384,27024,28384,15536,13800,14728,11912,9432,23568,22016,19296,21696,8232,6840,5316,4804,4940,37568,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,10064,6316,11424,11072,10592,12904,15712,13712,13424,12328,12920,10984,11592,11336,16320,16360,16560,15104,19568,29008,27328,23200,20224,14248,18592,23200,23680,27744,20848,17744,24992,18688,14624,17264,20016,18864,18624,12024,8944,25744,23328,14632,13176,8760,23104,22384,12504,10680,12272,11424,16576,21552,18688,18112,18240,14344,12224,14464,15520,16416,13768,15832,19440,16400,14864,11712,8768,8496,9808,17632,18992,14072,16000,11648,13808,11688,11112,9344,10488,8952,9744,9736,11272,7760,9464,18448,15688,14384,14496,17008,16688,13072,15464,18752,15904,19264,16184,16304,19440,20048,15368,14856,18768,15320,14552,8928,22592,11048,17584,22992,20928,9440,9040,17888,21968,20576,14984,17536,20608,14072,16168,11224,14720,20176,18288,8560,15832,13792,15456,15904,9920,20496,11080,14296,20448,14240,17728,15864,15488,14064,15840,15664,17904,14128,10568,12120,11320,11408,21280,22800,26384,16624,15872,18992,23008,10496,13344,13880,8040,7276,7372,7648,7072,7288,8188,6980,7200,8080,7212,10752,19408,19552,16480,11848,12808,14704,14088,13872,6876,7036,7136,4732,11688,15328,15072,18000,19760,15008,18640,18688,18592,19488,19552,17696,14720,18576,17792,14136,14624,8816,6788,6332,6180,4148,18048,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,12936,9920,16104,14880,13952,17344,17248,14120,12648,13760,13272,13728,14272,16304,18304,19344,19856,20528,20080,21984,20112,23648,22336,14488,22256,26736,22080,19760,17264,18672,21184,20992,18192,20688,23488,23760,21248,13768,6968,22736,16248,15888,14872,12528,19568,17632,13664,11080,11304,11832,12664,15520,14064,15384,15520,12976,11304,16328,18432,20848,19472,18592,27552,22880,19136,13512,13008,13712,16752,19664,21024,12960,11352,11120,14632,10696,10872,11584,11232,12192,11504,11752,12664,12208,11216,16368,11976,12288,14656,15080,15008,15760,18496,17632,20176,19408,17424,15440,15568,14824,14832,14256,20352,19040,23952,19216,29648,19696,29344,29328,23200,22800,22784,35904,39040,45888,40000,35296,34560,28800,30448,23024,36480,50560,41472,20480,31680,31184,30256,31680,23424,28352,23552,25776,31120,28928,32800,23600,21104,21216,23888,25488,24064,25056,23024,20384,20704,22800,37120,27040,32464,32688,29136,26640,29360,26368,24752,25760,21248,18256,16176,17888,14536,13896,13704,14792,15960,17168,15936,20032,21984,20624,21488,18880,18112,17200,19840,24128,17440,15232,15752,10432,17968,22864,22304,25920,25680,25088,31008,31104,10504,8112,7820,8248,8408,9680,9056,9512,8608,5868,4492,4476,5164,3150,11712,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]',
     ),
   ),
-};
+});
 
-export const multivecTileDataWithServerAggregation = {
+export const multivecTileDataWithServerAggregation = /** @type {const} */ ({
   dtype: 'float16',
   maxNonZero: 39840,
   minNonZero: 6172,
@@ -31,4 +30,4 @@ export const multivecTileDataWithServerAggregation = {
       '[8328,7776,13240,11104,8600,15760,13480,10560,10208,10840,9360,9672,11296,13392,18896,19568,20880,28048,33600,39840,30112,33952,27328,15144,27664,31376,31120,25104,21808,22352,30320,31744,25472,28976,30496,32464,23856,18256,8984,31968,25472,21104,17040,13016,23328,25312,16592,14440,12896,15544,19040,23152,14288,17392,19104,18992,12808,19072,25232,22432,16992,18912,28576,25104,22272,12888,16336,15568,16432,21056,21472,11760,12376,10528,14600,10072,9952,10032,11664,13224,13168,11656,11080,10712,11760,19120,15968,13776,12264,15080,15376,14216,19344,20448,18496,17936,12880,14472,17600,16544,15152,18016,20048,17776,20416,16184,21408,13920,21504,21072,21984,11056,12056,19376,29568,30448,24624,26832,27488,15288,16944,21872,26352,39328,33280,16608,23136,24000,25168,22992,20544,28176,22576,22544,26496,28048,22160,18304,17408,17632,19392,17360,22416,15960,15736,20048,15336,20976,30384,26096,31072,21360,21648,16288,21648,15856,17200,18080,15824,10136,10760,14792,11056,9360,9752,8336,10000,11056,10232,16080,14504,15656,15808,13912,13928,15720,17488,18128,14456,11240,12848,8052,10072,13960,10600,15992,20960,15304,24048,28832,21504,20784,20096,15112,13280,21472,22624,19280,20944,12384,8896,8060,8240,7088,34368,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16000,10864,19024,17216,17744,19744,21120,16928,17168,16656,15416,14384,15016,12936,14920,17248,22736,15032,17616,20960,19632,23984,21984,12368,20784,26080,18960,15832,17008,14216,17232,19760,14520,17216,20848,21808,18288,11232,7220,24592,17248,21360,20848,14304,20064,16312,15128,13864,14080,14608,14480,17232,16032,14168,14104,13960,12672,20080,21008,20768,18512,19472,22064,20880,20272,14720,13032,13496,13664,18080,17984,14776,15480,14592,17632,14952,14344,15144,14536,14592,13048,13424,15072,13536,14368,18448,15232,16024,17312,16592,14096,16848,21680,16784,16296,15368,14984,14880,17776,14592,14368,15008,23680,15360,15360,10776,21216,12160,21200,18992,14680,14328,14896,19168,24016,30336,21696,14624,21248,16544,19600,14976,19840,32528,25712,10904,14632,15304,20624,22496,14776,19248,13216,15960,19616,17984,24496,14472,13864,13872,15064,15528,15352,16752,15304,17904,15008,15888,28256,18272,20384,24576,17696,17632,13280,12760,16224,16432,12336,13712,13176,12744,12440,12712,12560,13112,12888,13640,13280,13544,12720,15096,12856,14400,13808,16560,20192,16032,11984,11696,12536,6172,14832,15504,14520,16048,15720,18256,20016,23520,13200,12688,13640,15064,15520,19888,16352,20928,18304,13224,14112,11560,11848,6828,26896,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,20368,13800,20656,18032,18016,19072,20720,19216,18224,17520,16592,15888,17136,17216,19744,20688,22368,16544,17792,22624,23152,23648,20416,18624,20784,21952,21008,21792,20640,18528,22944,20720,16832,18048,19936,20000,19792,13888,8840,22400,21008,20048,18960,14224,22704,21040,15768,15576,15936,16080,16672,19088,18464,19856,18848,17680,15288,18864,18928,20240,20032,20384,21168,21520,19344,16880,13712,13376,15304,19504,21312,17248,17600,17488,20128,16360,16528,16048,16160,14696,14568,16248,18608,15096,15920,22864,20848,20944,21680,21344,17984,16672,21760,19520,19744,20400,19536,19712,21504,20208,17888,17200,23168,17664,15536,11328,21616,14128,20992,21568,18160,15816,16144,22032,22624,22032,19696,18496,20496,20480,21264,17360,16960,21088,19344,12040,18912,18400,19712,19456,15816,19616,14752,17696,20912,19248,22832,19504,18608,17520,18640,20048,20880,20640,17808,17856,18000,17392,22800,22800,23472,21952,19936,21168,20496,17360,18800,18416,13320,14240,13720,14736,13560,14080,14560,14264,14528,15776,14288,16752,20704,20944,18688,17216,17312,18560,20560,18144,14304,15104,15040,10304,16216,21168,20816,21632,20544,19552,20080,21504,18736,18192,18880,19648,19312,20048,20096,17856,17536,14440,13080,12032,12392,8592,20688,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]',
     ),
   ),
-};
+});
